requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
streamlit>=1.40.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Gemini AI
google-genai>=0.2.0

# Data Processing
python-multipart>=0.0.20
python-dateutil>=2.8.2
pyyaml>=6.0.1
pandas>=2.2.0

# HTTP & Async
httpx>=0.26.0
aiofiles>=24.1.0

# Utilities
python-dotenv>=1.0.1
tenacity>=9.0.0

# Development
pytest>=8.3.0
pytest-asyncio>=0.24.0
black>=25.1.0
ruff>=0.9.0
